rbox-protocol: decouple command parsing from request processing
- Move parsing and env extraction from process_completed_request to rbox_server_get_request
- Add parse_valid flag to rbox_server_request_t
- Add rbox_server_request_parse_valid() accessor
- Update rbox_server_request_arg/argc fallbacks for invalid parse
- Remove parse-or-deny behavior; parse failures are non-fatal
- Fix unused variable warning in protocol_encoding.c
- Fix sign comparison warning in log.c

+++ rbox-protocol/src/log.c
@@ -60,7 +60,7 @@ void rbox_log_set_callback(rbox_log_callback_t cb) {
 
 __attribute__((visibility("default")))
 void rbox_log(rbox_log_level_t level, const char *file, int line, const char *fmt, ...) {
-    if (level > atomic_load_explicit(&g_log_level, memory_order_relaxed)) {
+    if ((int)level > (int)atomic_load_explicit(&g_log_level, memory_order_relaxed)) {
         return;
     }
